Work on implementing UserServise interfaces

+++ Squids-Movies-App/SquidsMovieApp.Data/Models/User.cs
@@ -47,7 +47,7 @@ namespace SquidsMovieApp.Data.Models
 
         public bool IsAdmin { get; set; }
 
-        public int MoneyBalance { get; set; }
+        public decimal MoneyBalance { get; set; }
 
         public virtual ICollection<Participant> LikedActors { get; set; }
         public virtual ICollection<Participant> LikedDirectors { get; set; }
